Module pica_record::matcher

Various matcher against record primitives.

Modules

field

Matcher that can be applied on a list of FieldRef.

subfield

Matcher that can be applied on a list of SubfieldRef.

Structs

MatcherOptions

Options and flags which can be used to configure a matcher.

ParseMatcherError

An error that can occur when parsing PICA+ matcher.

RecordMatcher

A matcher that matches against a RecordRef.

Enums

BooleanOp

Boolean Operators.

OccurrenceMatcher

A matcher that matches against a OccurrenceRef.

Quantifier

RelationalOp

Relational Operator

TagMatcher

A matcher that matches against a TagRef.