Enum pica_record::matcher::subfield::SubfieldMatcher

pub enum SubfieldMatcher {
    Singleton(SingletonMatcher),
    Group(Box<SubfieldMatcher>),
    Not(Box<SubfieldMatcher>),
    Composite {
        lhs: Box<SubfieldMatcher>,
        op: BooleanOp,
        rhs: Box<SubfieldMatcher>,
    },
}

A matcher that allows grouping, negation and connecting of singleton matcher.

Variants

Singleton(SingletonMatcher)

Group(Box<SubfieldMatcher>)

Not(Box<SubfieldMatcher>)

Composite

Fields

lhs: Box<SubfieldMatcher>

op: BooleanOp

rhs: Box<SubfieldMatcher>

Implementations

impl SubfieldMatcher

pub fn new(matcher: &str) -> Result<Self, ParseMatcherError>

Creates a new SubfieldMatcher.

Errors

This function fails if the given expression is not a valid subfield-matcher.

Example

use pica_record::matcher::subfield::SubfieldMatcher;

let _matcher = SubfieldMatcher::new("a == 'foo'")?;

pub fn is_match<'a>( &self, subfields: impl IntoIterator<Item = &'a SubfieldRef<'a>> + Clone, options: &MatcherOptions, ) -> bool

Returns true if the underlying matcher returns true.

Example

use pica_record::matcher::subfield::SubfieldMatcher;
use pica_record::matcher::MatcherOptions;
use pica_record::primitives::SubfieldRef;

let options = MatcherOptions::default();
let subfield = SubfieldRef::new('a', "foo")?;

let matcher = SubfieldMatcher::new("a =^ 'f' && (a =$ 'o')")?;
assert!(matcher.is_match(&subfield, &options));