Struct pica_record::primitives::SubfieldCode

pub struct SubfieldCode(/* private fields */);

A subfield code.

This type represents a PICA+ subfield code, which is a single ASCII alpha-numeric character.

Implementations

impl SubfieldCode

pub fn new(code: char) -> Result<Self, ParsePicaError>

Creates a new subfield code.

Errors

This function fails if the given code is not an ASCII alpha-numeric character.

Example

use pica_record::primitives::SubfieldCode;

let code = SubfieldCode::new('a')?;
assert_eq!(code, 'a');

pub fn from_unchecked<T: Into<char>>(code: T) -> Self

Creates a subfied code without checking for validity.

Safety

The caller must ensure that the given subfield code is valid.

Example

use pica_record::primitives::SubfieldCode;

let code = SubfieldCode::from_unchecked('a');
assert_eq!(code, 'a');

pub fn as_byte(&self) -> u8

Returns the subfield code as a byte (u8).

Example

use pica_record::primitives::SubfieldCode;

let code = SubfieldCode::new('a')?;
assert_eq!(code.as_byte(), b'a');