Struct pica_record::FieldRef

pub struct FieldRef<'a> { /* private fields */ }

An immutable PICA+ field.

Implementations

impl<'a> FieldRef<'a>

pub fn new<B: ?Sized + AsRef<[u8]>>( tag: &'a B, occurrence: Option<&'a B>, subfields: Vec<(char, &'a B)> ) -> Self

Create a new field.

Panics

This method panics if a parameter is invalid.

Example

use pica_record::{FieldRef, Tag};

fn example() -> anyhow::Result<()> {
    let field = FieldRef::new("012A", None, vec![('0', "abc")]);

    assert_eq!(field.tag(), b"012A");
    assert_eq!(field.subfields().len(), 1);
    assert!(field.occurrence().is_none());
    Ok(())
}

pub fn from_bytes(bytes: &'a [u8]) -> Result<Self, ParsePicaError>

Creates an immutable PICA+ field from a byte slice.

If an invalid field is given, an error is returned.

use pica_record::FieldRef;

fn example() -> anyhow::Result<()> {
    let field =
        FieldRef::from_bytes(b"003@ \x1f0123456789X\x1e").unwrap();
    Ok(())
}

pub fn tag(&self) -> &TagRef<'_>

Returns the tag of the field.

Example

use pica_record::{FieldRef, TagRef};

fn example() -> anyhow::Result<()> {
    let field =
        FieldRef::from_bytes(b"003@ \x1f0123456789X\x1e").unwrap();
    assert_eq!(field.tag(), &TagRef::new("003@"));

    Ok(())
}

pub fn occurrence(&self) -> Option<&OccurrenceRef<'_>>

Returns a reference to the occurrence of the field.

Example

use pica_record::{FieldRef, Occurrence};

fn example() -> anyhow::Result<()> {
    let field = FieldRef::new("012A", Some("01"), vec![]);
    let occurrence = field.occurrence().unwrap();
    assert_eq!(*occurrence, "01");

    Ok(())
}

pub fn subfields(&self) -> &Vec<SubfieldRef<'_>>

Returns the subfields of the field.

Example

use pica_record::FieldRef;

fn example() -> anyhow::Result<()> {
    let field = FieldRef::new(
        "012A",
        Some("01"),
        vec![('a', "b"), ('c', "d")],
    );

    assert_eq!(field.subfields().len(), 2);

    Ok(())
}

pub fn validate(&self) -> Result<(), Utf8Error>

Returns an std::str::Utf8Error if the field contains invalid UTF-8 data, otherwise the unit.

Example

use pica_record::FieldRef;

fn example() -> anyhow::Result<()> {
    let field = FieldRef::from_bytes(b"003@ \x1f0123\x1e")?;
    assert!(field.validate().is_ok());

    let field = FieldRef::from_bytes(b"003@ \x1f0\x00\x9F\x1e")?;
    ///     assert!(field.validate().is_err());
    Ok(())
}

pub fn write_to(&self, out: &mut impl Write) -> Result<()>

Write the field into the given writer.

Example

use std::io::Cursor;

use pica_record::FieldRef;

fn example() -> anyhow::Result<()> {
    let mut writer = Cursor::new(Vec::<u8>::new());
    let field = FieldRef::from_bytes(b"012A/01 \x1fab\x1fcd\x1e")?;
    field.write_to(&mut writer);

    Ok(())
}

pub fn level(&self) -> Level

Returns the level of the field.

Example

use pica_record::{FieldRef, Level};

fn example() -> anyhow::Result<()> {
    let field = FieldRef::from_bytes(b"012A/01 \x1fab\x1fcd\x1e")?;
    assert_eq!(field.level(), Level::Main);

    Ok(())
}