Struct pica_record::RecordRef

pub struct RecordRef<'a>(/* private fields */);

An immutable PICA+ record.

Implementations

impl<'a> RecordRef<'a>

pub fn new<B>(fields: Vec<(&'a B, Option<&'a B>, Vec<(char, &'a B)>)>) -> Self
where B: ?Sized + AsRef<[u8]>,

Create a new immutable record.

Panics

This method panics if a parameter is invalid.

Example

use pica_record::RecordRef;

fn example() -> anyhow::Result<()> {
    let record =
        RecordRef::new(vec![("003@", None, vec![('0', "abc")])]);
    assert_eq!(record.iter().len(), 1);

    Ok(())
}

pub fn from_bytes(bytes: &'a [u8]) -> Result<Self, ParsePicaError>

Creates an PICA+ record from a byte slice.

If an invalid record is given, an error is returned.

Example

use pica_record::RecordRef;

fn example() -> anyhow::Result<()> {
    let record = RecordRef::from_bytes(b"003@ \x1f0abc\x1e\n");
    assert_eq!(record.iter().len(), 1);

    Ok(())
}

pub fn is_empty(&self) -> bool

Returns true if the record contains no fields, otherwise false.

Example

use pica_record::RecordRef;

fn example() -> anyhow::Result<()> {
    let record = RecordRef::from_bytes(b"002@ \x1f0Oaf\x1e\n")?;
    assert!(!record.is_empty());

    Ok(())
}

pub fn iter(&self) -> Iter<'_, FieldRef<'_>>

Returns an iterator over the fields of the record.

Panics

This method panics if a parameter is invalid.

Example

use pica_record::RecordRef;

fn example() -> anyhow::Result<()> {
    let record = RecordRef::new(vec![
        ("003@", None, vec![('0', "123456789X")]),
        ("002@", None, vec![('0', "Oaf")]),
    ]);

    assert_eq!(record.iter().len(), 2);
    Ok(())
}

pub fn retain<F: FnMut(&FieldRef<'_>) -> bool>(&mut self, f: F)

Retains only the fields specified by the predicate.

Example

use pica_record::{RecordRef, TagRef};

fn example() -> anyhow::Result<()> {
    let mut record = RecordRef::new(vec![
        ("003@", None, vec![('0', "123456789X")]),
        ("002@", None, vec![('0', "Oaf")]),
    ]);

    record.retain(|field| field.tag() == &TagRef::new("003@"));
    assert_eq!(record.iter().len(), 1);

    Ok(())
}

pub fn validate(&self) -> Result<(), Utf8Error>

Returns an std::str::Utf8Error if the record contains invalid UTF-8 data, otherwise the unit.

Example

use pica_record::RecordRef;

fn example() -> anyhow::Result<()> {
    let record = RecordRef::from_bytes(b"003@ \x1f0a\x1e\n")?;
    assert!(record.validate().is_ok());
    Ok(())
}

pub fn write_to(&self, out: &mut impl Write) -> Result<()>

Write the record into the given writer.

Example

use std::io::Cursor;

use pica_record::RecordRef;

fn example() -> anyhow::Result<()> {
    let mut writer = Cursor::new(Vec::<u8>::new());
    let record = RecordRef::from_bytes(b"003@ \x1f0a\x1e\n")?;
    record.write_to(&mut writer);

    Ok(())
}