Struct pica_record::Subfield

pub struct Subfield<T: AsRef<[u8]>> { /* private fields */ }

A PICA+ subfield.

Implementations

impl<T: AsRef<[u8]>> Subfield<T>

pub fn code(&self) -> char

Returns the code of the subfield.

Example

use pica_record::SubfieldRef;

fn example() -> anyhow::Result<()> {
    let subfield = SubfieldRef::new('0', "0123456789X");
    assert_eq!(subfield.code(), '0');
    Ok(())
}

pub fn value(&self) -> &T

Returns the value of the subfield.

Example

use pica_record::SubfieldRef;

fn example() -> anyhow::Result<()> {
    let subfield = SubfieldRef::new('0', "123456789X");
    assert_eq!(subfield.value(), &"123456789X".as_bytes());

    Ok(())
}

impl<'a, T: AsRef<[u8]> + From<&'a BStr> + Display> Subfield<T>

pub fn new(code: char, value: impl Into<T>) -> Self

Create a new subfield.

Panics

This method panics if the subfield code or the value is invalid.

Example

use pica_record::SubfieldRef;

fn example() -> anyhow::Result<()> {
    let subfield = SubfieldRef::new('a', "bcd");
    assert_eq!(subfield.code(), 'a');
    assert_eq!(subfield.value(), &"bcd".as_bytes());
    Ok(())
}

pub fn from_bytes(data: &'a [u8]) -> Result<Self, ParsePicaError>

Creates an immutable PICA+ subfield from a byte slice.

If an invalid subfield is given, an error is returned.

Example

use pica_record::SubfieldRef;

fn example() -> anyhow::Result<()> {
    let subfield = SubfieldRef::from_bytes(b"\x1f0123456789X")?;
    assert_eq!(subfield.code(), '0');
    assert_eq!(subfield.value(), &"123456789X".as_bytes());
    Ok(())
}

pub fn is_empty(&self) -> bool

Returns true if the subfield value is empty.

Example

use pica_record::SubfieldRef;

fn example() -> anyhow::Result<()> {
    let subfield = SubfieldRef::new('0', "abc");
    assert!(!subfield.is_empty());

    let subfield = SubfieldRef::new('0', "");
    assert!(subfield.is_empty());
    Ok(())
}

pub fn validate(&self) -> Result<(), Utf8Error>

Returns an std::str::Utf8Error if the subfield value contains invalid UTF-8 data, otherwise the unit.

Example

use pica_record::SubfieldRef;

fn example() -> anyhow::Result<()> {
    let subfield = SubfieldRef::new('0', "123456789X");
    assert_eq!(subfield.validate().is_ok(), true);

    let subfield =
        SubfieldRef::from_bytes(&[b'\x1f', b'0', 0, 159])?;
    assert_eq!(subfield.validate().is_err(), true);

    Ok(())
}

pub fn write_to(&self, out: &mut impl Write) -> Result<()>

Write the subfield into the given writer.

Example

use std::io::Cursor;

use pica_record::SubfieldRef;

fn example() -> anyhow::Result<()> {
    let mut writer = Cursor::new(Vec::<u8>::new());
    let subfield = SubfieldRef::new('0', "123456789X");
    subfield.write_to(&mut writer);

    Ok(())
}

impl<'a> Subfield<&'a BStr>

pub fn into_owned(self) -> SubfieldMut

Converts the immutable subfield into its mutable counterpart by consuming the source.

Example

use pica_record::SubfieldRef;

fn example() -> anyhow::Result<()> {
    let subfield =
        SubfieldRef::new('0', "0123456789X").into_owned();
    assert_eq!(subfield.value(), "0123456789X");
    Ok(())
}

pub fn to_owned(&self) -> SubfieldMut

Converts the immutable subfield into its mutable counterpart.

Example

use pica_record::SubfieldRef;

fn example() -> anyhow::Result<()> {
    let subfield = SubfieldRef::new('0', "0123456789X").to_owned();
    assert_eq!(subfield.value(), "0123456789X");
    Ok(())
}