Crate pica_record

This crate provides the low-level primitives to work with bibliographic records encoded in PICA+.

There exists an immutable and a mutable variant for each primitive. The immutable variant is used to parse the corresponding component of a record, without owning the data. This type is mostly a wrapper of the underlying data (byte slices). On the other hand there is also a mutable variant, which is used in upstream crates. This variant owns it’s data.

This crate also provides two higher-level data structure to work with records: ByteRecord and StringRecord. The first type is a wrapper of a RecordRef and provides more functions to work with records as well as an mechanism to cache the complete byte sequence of the original record. This improves the performance when the record is written back to a stream. It is important to note that a ByteRecord might contain invalid UTF-8 data. When it’s important to guarantee valid UTF-8 data, use a StringRecord instead.

Finally, the io module provides utilities for reading and writing PICA+ records and the parser module exposes the internal parser combinators, which are used in upstream crates (matcher, select).

Modules

io

Utilities for reading and writing PICA+ records.

parser

Parsers recognizing low-level primitives (e.g. subfield codes).

Structs

ByteRecord

A PICA+ record, that may contain invalid UTF-8 data.

Field

A mutable PICA+ field.

FieldRef

An immutable PICA+ field.

Occurrence

A mutable PICA+ occurrence.

OccurrenceRef

An immutable PICA+ occurrence.

ParseLevelError

An error that can occur when parsing PICA+ level.

Record

An immutable PICA+ record.

RecordRef

An immutable PICA+ record.

StringRecord

A PICA+ record, that guarantees valid UTF-8 data.

Subfield

A mutable PICA+ subfield.

SubfieldRef

An immutable PICA+ subfield.

Tag

A mutable PICA+ tag.

TagRef

An immutable PICA+ tag.

Enums

Level

The level (main, local, copy) of a field (or tag).

ParsePicaError

An error that can occur when parsing PICA+ records.