Trait pica_record::io::BufReadExt

pub trait BufReadExt: BufRead {
    // Provided method
    fn for_pica_record<F>(&mut self, f: F) -> Result<(), ReadPicaError>
       where F: FnMut(Result<ByteRecord<'_>, ParsePicaError>) -> Result<bool, ReadPicaError> { ... }
}

An extension of BufRead which provides a convenience API for reading ByteRecords.

Provided Methods

fn for_pica_record<F>(&mut self, f: F) -> Result<(), ReadPicaError>where F: FnMut(Result<ByteRecord<'_>, ParsePicaError>) -> Result<bool, ReadPicaError>,

Executes the given closure on each parsed line in the underlying reader.

If the underlying reader or the closure returns an error, then the iteration stops and the error is returned. If the closure returns false the iteration is stopped and no error is returned.

Example

use std::io::{Cursor, Seek};

use pica_record::io::BufReadExt;

fn example() -> anyhow::Result<()> {
    let mut reader =
        Cursor::new(b"003@ \x1f0abc\x1e\n003@ \x1f0def\x1e\n");

    // iterate over all records
    let mut count = 0;
    reader.for_pica_record(|result| {
        let _record = result?;
        count += 1;
        Ok(true)
    })?;

    assert_eq!(count, 2);

    // stop iteration after first record
    reader.rewind()?;
    count = 0;
    reader.for_pica_record(|result| {
        let _record = result?;
        count += 1;
        Ok(false)
    })?;

    assert_eq!(count, 1);

    Ok(())
}